--- a/src/python_osw_validation/__init__.py
+++ b/src/python_osw_validation/__init__.py
@@ -1,9 +1,9 @@
 import os
 import gc
 import json
-import jsonschema
-from typing import Dict, Any, Optional, List
+import jsonschema_rs
 import geopandas as gpd
+from typing import Dict, Any, Optional, List
 from .zipfile_handler import ZipFileHandler
 from .extracted_data_validator import ExtractedDataValidator, OSW_DATASET_FILES
 from .version import __version__
@@ -37,7 +37,7 @@ class OSWValidation:
         except Exception as e:
             self.errors.append(f'Invalid or missing schema file: {e}')
             raise Exception(f'Invalid or missing schema file: {e}')
-        
+
     def are_ids_unique(self, gdf):
         """Check for duplicate values in the _id field"""
         duplicates = gdf[gdf.duplicated('_id', keep=False)]['_id'].unique()
@@ -62,6 +62,7 @@ class OSWValidation:
                 self.errors.append(validator.error)
                 return ValidationResult(False, self.errors)
 
+            # Validate schema for each file
             for file in validator.files:
                 file_path = os.path.join(file)
                 if not self.validate_osw_errors(file_path=str(file_path), max_errors=max_errors):
@@ -70,82 +71,56 @@ class OSWValidation:
             if self.errors:
                 return ValidationResult(False, self.errors)
 
-            # Validate data integrity
+            # Validate data integrity, freeing up memory after processing each file
             for file in validator.files:
                 file_path = os.path.join(file)
-                osw_file = next((osw_file_any for osw_file_any in OSW_DATASET_FILES.keys() if osw_file_any in file_path), '')
+                osw_file = next(
+                    (osw_file_any for osw_file_any in OSW_DATASET_FILES.keys() if osw_file_any in file_path), ''
+                )
                 OSW_DATASET[osw_file] = gpd.read_file(file_path)
 
-            # Are all id's unique in each file? No need to check uniqueness across files yet since we do not have a global OSW ID format yet
-            for osw_file in OSW_DATASET:
+                # Validate uniqueness of _id fields in each file
                 is_valid, duplicates = self.are_ids_unique(OSW_DATASET[osw_file])
                 if not is_valid:
                     self.errors.append(f"Duplicate _id's found in {osw_file} : {duplicates}")
 
-            # Create sets of node id's and foreign keys to be used in validation
-            if 'nodes' in OSW_DATASET:
-                node_ids = set(OSW_DATASET['nodes']['_id'])
-            else:
-                node_ids = set()
+                # Geometry validation
+                invalid_geojson = OSW_DATASET[osw_file][
+                    (OSW_DATASET[osw_file].geometry.type != OSW_DATASET_FILES[osw_file]['geometry']) |
+                    (~OSW_DATASET[osw_file].is_valid)
+                    ]
+                if len(invalid_geojson) > 0:
+                    self.errors.append(f"Invalid {osw_file} geometries: {invalid_geojson['_id'].tolist()}")
+                del invalid_geojson  # Release memory after validation of each file
 
-            if 'edges' in OSW_DATASET:
-                node_ids_edges_u = set(OSW_DATASET['edges']['_u_id'])
-                node_ids_edges_v = set(OSW_DATASET['edges']['_v_id'])
-            else:
-                node_ids_edges_u = set()
-                node_ids_edges_v = set()
+                # Delete the GeoDataFrame for this file after processing
+                del OSW_DATASET[osw_file]
+                gc.collect()  # Ensure memory is freed
 
-            if 'zones' in OSW_DATASET:
-                node_ids_zones_w = set([item for sublist in OSW_DATASET['zones']['_w_id'] for item in sublist])
-            else:
-                node_ids_zones_w = set()
-
-            # Do all node references in _u_id exist in nodes?
-            unmatched = node_ids_edges_u - node_ids
-            is_valid = len(unmatched) == 0
-            if not is_valid:
-                self.errors.append(f"All _u_id's in edges should be part of _id's mentioned in nodes, _u_id's not in nodes are: {unmatched}")
-
-            # Do all node references in _v_id exist in nodes?
-            unmatched = node_ids_edges_v - node_ids
-            is_valid = len(unmatched) == 0
-            if not is_valid:
-                self.errors.append(f"All _v_id's in edges should be part of _id's mentioned in nodes, _v_id's not in nodes are: {unmatched}")
-
-            # Do all node references in _w_id exist in nodes?
-            unmatched = node_ids_zones_w - node_ids
-            is_valid = len(unmatched) == 0
-            if not is_valid:
-                self.errors.append(f"All _w_id's in zones should be part of _id's mentioned in nodes, _w_id's not in nodes are: {unmatched}")
-
-            # Geometry validation: check geometry type in each file and test if coordinates make a shape that is reasonable geometric shape according to the Simple Feature Access standard
-            for osw_file in OSW_DATASET:
-                invalid_geojson = OSW_DATASET[osw_file][(OSW_DATASET[osw_file].geometry.type != OSW_DATASET_FILES[osw_file]['geometry']) | (OSW_DATASET[osw_file].is_valid == False)]
-                is_valid = len(invalid_geojson) == 0
-                if not is_valid:
-                    self.errors.append(f"Invalid {osw_file} geometries found, id's of invalid geometries: {set(invalid_geojson['_id'])}")
-
-            # Validate OSW external extensions
+            # Validate OSW external extensions and free memory
             for file in validator.externalExtensions:
                 file_path = os.path.join(file)
-                extensionFile = gpd.read_file(file_path)
-                invalid_geojson = extensionFile[extensionFile.is_valid == False]
-                is_valid = len(invalid_geojson) == 0
-                if not is_valid:
-                    self.errors.append(f"Invalid geometries found in extension file {file}, list of invalid geometries: {invalid_geojson.to_json()}")
+                extension_file = gpd.read_file(file_path)
+                invalid_geojson = extension_file[~extension_file.is_valid]
+                if len(invalid_geojson) > 0:
+                    self.errors.append(f"Invalid geometries in {file}: {invalid_geojson['_id'].tolist()}")
+                del extension_file, invalid_geojson  # Free memory after processing each extension
+                gc.collect()
 
             if self.errors:
                 return ValidationResult(False, self.errors)
             else:
                 return ValidationResult(True)
+
         except Exception as e:
             self.errors.append(f'Unable to validate: {e}')
             return ValidationResult(False, self.errors)
+
         finally:
-            del OSW_DATASET
+            del OSW_DATASET  # Release memory from the OSW_DATASET GeoDataFrames
             if zip_handler:
                 zip_handler.remove_extracted_files()
-            gc.collect()
+            gc.collect()  # Ensure all unused memory is released
 
     def load_osw_file(self, graph_geojson_path: str) -> Dict[str, Any]:
         """Load OSW Data"""
@@ -155,11 +130,19 @@ class OSWValidation:
     def validate_osw_errors(self, file_path: str, max_errors: int) -> bool:
         """Validate OSW Data against the schema and process all errors"""
         geojson_data = self.load_osw_file(file_path)
-        validator = jsonschema.Draft7Validator(self.load_osw_schema(self.schema_file_path))
+        validator = jsonschema_rs.Draft7Validator(self.load_osw_schema(self.schema_file_path))
 
+        error_count = 0
         for error in validator.iter_errors(geojson_data):
-            self.errors.append(f'Validation error: {error.message}')
-            if len(self.errors) >= max_errors:
-                return False
+            if error_count < max_errors:
+                self.errors.append(f'Validation error: {error.message}')
+                error_count += 1
+            else:
+                break
+
+        # If we've hit the max_errors, release memory associated with errors
+        if len(self.errors) >= max_errors:
+            del self.errors[:]  # Clear the error list to free up memory
+            gc.collect()
 
         return len(self.errors) < max_errors